+++ R/reexport.R
@@ -2,10 +2,6 @@
 #' @export
 generics::tidy
 
-#' @importFrom magrittr %>%
-#' @export
-magrittr::`%>%`
-
 #' @importFrom parsnip multi_predict
 #' @export
 parsnip::multi_predict