+++ tests/testthat/test-reg-spls.R
@@ -10,9 +10,9 @@ test_that("mixOmics multivariate model fitting", {
 
   expect_error(
     parsnip_spls_multi <-
-      parsnip::pls(num_comp = 3, predictor_prop = 1) %>%
-      set_engine("mixOmics") %>%
-      set_mode("regression") %>%
+      parsnip::pls(num_comp = 3, predictor_prop = 1) |>
+      set_engine("mixOmics") |>
+      set_mode("regression") |>
       fit_xy(x = meats_x_tr, y = meats_y_tr),
     regexp = NA
   )
@@ -74,9 +74,9 @@ test_that("mixOmics univariate model fitting", {
 
   expect_error(
     parsnip_spls_uni <-
-      parsnip::pls(num_comp = 3, predictor_prop = 1) %>%
-      set_engine("mixOmics") %>%
-      set_mode("regression") %>%
+      parsnip::pls(num_comp = 3, predictor_prop = 1) |>
+      set_engine("mixOmics") |>
+      set_mode("regression") |>
       fit_xy(x = meats_x_tr, y = meats_y_tr[[1]]),
     regexp = NA
   )